requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
backports.zoneinfo==0.2.1; python_version < "3.9"
boto3==1.37.38
botocore==1.37.38
certifi==2025.10.5
charset-normalizer==2.0.12
distlib==0.3.4
Django==3.2.25
django-cors-headers==4.4.0
django-redis==5.2.0
django-storages==1.14.6
djangorestframework==3.13.1
drf-spectacular==0.28.0
filelock==3.6.0
idna==3.11
importlib_resources==6.4.5
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
pillow==10.4.0
pkgutil_resolve_name==1.3.10
platformdirs==2.5.1
psycopg2==2.8.6
python-dateutil==2.9.0.post0
python-dotenv==0.19.2
pytz==2025.2
PyYAML==6.0.3
redis==5.0.1
referencing==0.35.1
requests==2.27.1
rpds-py==0.20.1
s3transfer==0.11.5
six==1.16.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==1.26.20
virtualenv==20.13.3
zipp==3.20.2
